+++ src/django_agent_trust/context_processors.py
@@ -1,6 +1,3 @@
-from __future__ import absolute_import, division, print_function, unicode_literals
-
-
 def agent(request):
     """
     A context processor that sets the template context variable ``agent`` to