+++ src/django_agent_trust/decorators.py
@@ -1,5 +1,3 @@
-from __future__ import absolute_import, division, print_function, unicode_literals
-
 from functools import wraps
 
 from django.contrib.auth.decorators import user_passes_test