+++ src/django_agent_trust/middleware.py
@@ -1,5 +1,3 @@
-from __future__ import absolute_import, division, print_function, unicode_literals
-
 from base64 import b64decode, b64encode
 from datetime import datetime
 from hashlib import md5